+++ include/bit.cl
@@ -1,4 +1,4 @@
-__kernel void bitonic_sort(__global double *data, const int n) {
+__kernel void bitonic_sort(__global float *data, const int n) {
     int index = get_global_id(0);
 
     for (int k = 2; k <= n; k *= 2) {
@@ -8,13 +8,13 @@ __kernel void bitonic_sort(__global double *data, const int n) {
             if (ixj > index) {
                 if ((index / k) % 2 == 0) {
                     if (data[index] > data[ixj]) {
-                        int temp = data[index];
+                        float temp = data[index];
                         data[index] = data[ixj];
                         data[ixj] = temp;
                     }
                 } else {
                     if (data[index] < data[ixj]) {
-                        int temp = data[index];
+                        float temp = data[index];
                         data[index] = data[ixj];
                         data[ixj] = temp;
                     }
